Update install_company_canon_printer_drivers.sh
After further testing the version I provided previously was not installing the printer drivers if the drivers were not installed at all, likely a formatting/globbing issue (even though the version checking with the driver present was working fine). So I made a further modification that uses if/elif rather than a series of if statements and replaced the previous formatting. I have tested this on 10.12.5 with a machine with no drivers installed, a newer version installed, and an older version installed, and all of the behavior is as expected.